requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.39.9
botocore==1.39.9
celery==5.5.3
channels==4.2.2
channels_redis==4.2.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
flower==2.0.1
humanize==4.12.3
jmespath==1.0.1
kombu==5.5.4
msgpack==1.1.1
packaging==25.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.2.0
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
tornado==6.5.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13